Implementing factory model for building halo profile objects
